+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- More specific error messages are shown when a workflow run / synthesis gets interrupted.
+
 ## [1.1.0] - 2021-05-28
 
 ### Added