+++ CHANGELOG.md
@@ -8,12 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Component to grant or revoke administrator privileges to / from users.
+- Grant or revoke administrator privileges to / from users in the administrator page.
+- Manage access to a domain in the domain edit page.
 - Endpoint for changing the administrator status of users.
 - (Admin only) endpoint for getting a list of all users.
 - Endpoint to get all users with certain access to a domain.
 - Endpoint to set a user's access to a domain.
-- Component to manage access to a domain.
 
 ### Changed