+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add component to grant or revoke administrator privileges to / from users.
-- Add endpoint for changing the administrator status of users.
-- Add (admin only) endpoint for getting a list of all users.
+- Component to grant or revoke administrator privileges to / from users.
+- Endpoint for changing the administrator status of users.
+- (Admin only) endpoint for getting a list of all users.
+- Endpoint to get all users with certain access to a domain.
+- Endpoint to set a user's access to a domain.
 
 ### Changed